Added the RecentUserLocationDTO to the client and processed the incoming message using yajl_JSON and the DictionaryModelMapper. I reused the MeepNotificationManager to notify the LiveMapView when there is a location update coming from the WebSocketManager.
Added an OtherUserAnnotation object to represent all other users in the map apart from the current user and all users inherit a base class which conforms to the MKAnnotation protocol. When a new annotation is added to the map the map zooms to view all the annotations using an MKMapVew category that adds the method zoomToFitAnnotations.

Also added a green marble image to represent all the other users (the current user is a blue marble). At a later date these will be replaced with other images and/or a custom image provided by the user.