Modified the LiveMapViewController so that it can be initialised with a meeting. When a meeting is set, it only shows that meetings annotations. This is so that the view can be loaded from the main screen and also from within the MeetingDetailView.
Modified the alertMeSlider so that it is preset with the value that the user previously set.

The MeetingsDetailView is now managed by the MenuNavigationController so that it can be pushed onto the controller stack when a local notification wants to display the Live Map view for a specific meeting.

The User and Logout manager classes are now managed inside the MenuNavigationController (as opposed to the view controller).

All 'show' methods in the MenuNavigationController now have 'animated' as an argument.