The current user is now loaded after login (as it was previously) and when the application becomes active and the last time it retrieved the current user was more than a minute ago. Whilst the current user is being requested a HUD is shown, preventing the user from using the app. If the request fails then they only have the option to retry.
The same hud is displayed (with a different label) when a user requests to logout and they have the option to retry if the request fails again. The UIAlertView for the retry request is different when logging out and getting the current user as when logging out, they have the option to dismiss the retry whereas when getting the current user they don't.

The ability to retry a request has been added to the RequestManager class (which all RequestManagers inherit) and so the method is available for all its children. When a request is retried the ASIHTTPRequest class does not return if there is zero network reachability e.g. in "Airplane mode" and so a check for this has been added.